Moved `Namespace.find_by_name` to an outerjoin
Having the address class use `find_by_name`, cleaned up quite a bit
of context code.  Moving to an outer join allows us to return the namespace
even when no childeren exist.
